requirements.txt
cython
dash
git+https://github.com/lucasb-eyer/pydensecrf.git
matplotlib
numpy
Pillow
rasterio
pyproj
shapely
requests
tensorflow==2.4